+++ src/Processes.py
@@ -40,6 +40,9 @@ def processes_gui_func():
     searchentry2101 = builder.get_object('searchentry2101')
     button2101 = builder.get_object('button2101')
 
+    global initial_already_run
+    initial_already_run = 0
+
 
     # Processes tab GUI functions
     # --------------------------------- Called for running code/functions when button is pressed on the treeview ---------------------------------
@@ -205,6 +208,9 @@ def processes_initial_func():
 
     global filter_column
     filter_column = processes_data_list[0][2] - 1                                             # Search filter is "Process Name". "-1" is used because "processes_data_list" has internal column count and it has to be converted to Python index. For example, if there are 3 internal columns but index is 2 for the last internal column number for the relevant treeview column.
+
+    global initial_already_run
+    initial_already_run = 1
 
 
 # ----------------------------------- Processes - Get Process Data Function (gets processes data, adds into treeview and updates it) -----------------------------------